requirements.txt
appnope==0.1.0
decorator==4.1.2
Django==1.11.4
djangorestframework==3.6.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
